backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pymysql==1.1.0
python-dotenv==1.0.0
sentence-transformers==2.2.2
pandas==2.1.3
numpy==1.26.2
fastapi-cache2[redis]==0.2.1
redis==5.0.1
pydantic==2.5.0
